requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# PDF processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Computer Vision and Deep Learning
opencv-python>=4.8.0
torch>=2.0.0
transformers>=4.30.0

# Text processing
rapidfuzz>=3.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
gunicorn>=21.2.0

# Database (Prisma)
prisma==0.13.1

# Environment variables
python-dotenv>=1.0.0

# Auth (JWT)
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# LLM (OpenRouter)
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Gemini fallback
google-generativeai>=0.8.3
